Removed scss folder, created variable for max width, fixed font import issues.
