+++ gulp/release-tasks/assets-core.js
@@ -10,7 +10,6 @@ module.exports = function (gulp, plugins, config) {
       // Test if the element is set to deploy this component
       if (config.output[element].assetsCore === true) {
         return gulp.src(target, { dot: true })
-          .pipe(plugins.include({ hardFail: true }))
           .on('error', console.log)
           .pipe(gulp.dest(`${config.basepath.release}/${config.output[element].dest}/assets/${config.versionName}/`));
       } else {
