+++ src/assets/_project/_blocks/layout/_layout-scaffolding.scss
@@ -2,7 +2,7 @@
 // Primary styles
 
 .container-fluid.qg-site-width{
-  max-width: 1327px; // old SWE width
+  max-width: $qg-max-container-width; // old SWE width
 }
 
 #qg-primary-content {
