+++ src/assets/_project/_blocks/scss-general/_qg-variables.scss
@@ -74,6 +74,8 @@ $qg-footer-bg-alt-color: $qg-dark-blue;
 // ###################################
 // Rulers & spacing
 
+$qg-max-container-width: 1327px; // old SWE width
+
 $qg-spacing: 50px; // Note: Used by $grid-gutter-width
 $qg-spacing-large: $qg-spacing * 1.5;
 $qg-spacing-small: $qg-spacing / 2;
