+++ tests/reports/eslint/report.html
@@ -275,13 +275,13 @@
                 </tr>
                 <tr>
                   <td>Clean files</td>
-                  <td>32</td>
+                  <td>33</td>
                 </tr>
               </tbody>
               <tfoot>
                 <tr>
                   <td>Total files linted</td>
-                  <td>37</td>
+                  <td>38</td>
                 </tr>
               </tfoot>
           </table>
@@ -351,7 +351,7 @@
                 <td colspan="3">
                   <table class="details-table">
                     <tbody><tr class="msg-warning"><td class="location-col">
-                              86:23
+                              87:23
                           </td>
                           <td class="severity-col">Warning</td>
                           <td class="message-col">Unexpected &#x27;fixme&#x27; comment.</td>
@@ -373,20 +373,6 @@
       </tr></tbody>
                   </table>
                 </td>
-              </tr><tr class="warning"><td>            <a class="toggle-link" href="javascript:;" onclick="toggleDetails(this)">C:\git\glue-template\src\assets\_project\_blocks\layout\footer\language-scroller.js</a></td>
-              <td>0</td>
-              <td>1</td></tr><tr class="details-row hidden">
-                <td colspan="3">
-                  <table class="details-table">
-                    <tbody><tr class="msg-warning"><td class="location-col">
-                              34:64
-                          </td>
-                          <td class="severity-col">Warning</td>
-                          <td class="message-col">Missing trailing comma.</td>
-                            <td class="rule-col"><a href="http://eslint.org/docs/rules/comma-dangle">comma-dangle</a></td>
-      </tr></tbody>
-                  </table>
-                </td>
               </tr><tr class="warning"><td>            <a class="toggle-link" href="javascript:;" onclick="toggleDetails(this)">C:\git\glue-template\gulp\release-tasks\ssi-to-static.js</a></td>
               <td>0</td>
               <td>1</td></tr><tr class="details-row hidden">
@@ -398,6 +384,20 @@
                           <td class="severity-col">Warning</td>
                           <td class="message-col">Unexpected &#x27;todo&#x27; comment.</td>
                             <td class="rule-col"><a href="http://eslint.org/docs/rules/no-warning-comments">no-warning-comments</a></td>
+      </tr></tbody>
+                  </table>
+                </td>
+              </tr><tr class="warning"><td>            <a class="toggle-link" href="javascript:;" onclick="toggleDetails(this)">C:\git\glue-template\src\assets\_project\_blocks\layout\footer\language-scroller.js</a></td>
+              <td>0</td>
+              <td>1</td></tr><tr class="details-row hidden">
+                <td colspan="3">
+                  <table class="details-table">
+                    <tbody><tr class="msg-warning"><td class="location-col">
+                              34:64
+                          </td>
+                          <td class="severity-col">Warning</td>
+                          <td class="message-col">Missing trailing comma.</td>
+                            <td class="rule-col"><a href="http://eslint.org/docs/rules/comma-dangle">comma-dangle</a></td>
       </tr></tbody>
                   </table>
                 </td>
@@ -426,6 +426,8 @@
               <td>0</td></tr><tr class="success"><td>C:\git\glue-template\gulp\release-tasks\assets-includes-cdn.js</td>
               <td>0</td>
               <td>0</td></tr><tr class="success"><td>C:\git\glue-template\gulp\release-tasks\assets-includes-local.js</td>
+              <td>0</td>
+              <td>0</td></tr><tr class="success"><td>C:\git\glue-template\gulp\release-tasks\copy-element-html.js</td>
               <td>0</td>
               <td>0</td></tr><tr class="success"><td>C:\git\glue-template\gulp\release-tasks\copy-element.js</td>
               <td>0</td>
